refactored css in various files
